Fix register screen layout issue

+++ Food/Model/FoodPrice.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct FoodPriceCount {
 //    var foodPriceCount: Food?
-        var quantity: Decimal
-        var foodPrice: Decimal
+        var quantity: Double
+        var foodPrice: Double
     
 }
